backend/requirements.txt
asgiref==3.8.1
async-timeout==5.0.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.0
defusedxml==0.8.0rc2
Django==4.2.18
django-appconf==1.0.6
django-cors-headers==4.6.0
django-filter==24.3
django-imagekit==5.0.0
django-redis==5.4.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
djoser==2.3.1
idna==3.10
node==1.2.2
nodejs==0.1.1
npm==0.1.1
oauthlib==3.2.2
odict==1.9.0
optional-django==0.1.0
pilkit==3.0
pillow==11.1.0
plumber==1.7
pycparser==2.22
PyJWT==2.10.1
python-decouple==3.8
python3-openid==3.2.0
psycopg2-binary==2.9.6
redis==5.2.1
requests==2.32.3
requests-oauthlib==2.0.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
whitenoise==6.8.2
zope.component==6.0
zope.deferredimport==5.0
zope.deprecation==5.1
zope.event==5.0
zope.hookable==7.0
zope.interface==7.2
zope.lifecycleevent==5.0
zope.proxy==6.1
gunicorn==20.1.0